+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - Removed mobile hamburger/overlay UI; panels follow standard responsive layout at all screen sizes
+- Mobile devices now see a notice that mobile support is temporarily unavailable
 
 ## [2026-03-03]